[Ref] sounddev: Silence 64bit warning.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@4770 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ sounddev/SoundDevice.h
@@ -251,7 +251,7 @@ public:
 	
 	uint32 GetNumHostChannels() const
 	{
-		return ChannelToDeviceChannel.size();
+		return static_cast<uint32>(ChannelToDeviceChannel.size());
 	}
 	
 	// Get the number of required device channels for this mapping. Derived from the maximum mapped-to channel number.
